Introduce StreamListener to optimize boosting
The changes introduced in this commit replace the former MastoBoost class with the Listener class implementing StreamListener. Instead of continuously polling for new toots containing specific hashtags, the application now listens for public and unlisted toots and processes them on the fly. This approach reduces complexity, improves performance and should be less resource-intensive. Updates to Dockerfile, .dockerignore, options.py, README, and cli.py are made accordingly.