Upgrade GitHub Actions versions in workflow
The difference in this commit updates the version of multiple GitHub Actions used in the workflow. This includes the actions for checking out code, setting up Docker Buildx, logging into Docker, and building and pushing Docker images. By upgrading these dependencies we are ensuring the project benefits from the latest features and bug fixes.